code: Create event listener to open and show card icon
The starter code gives instructions to create a function open() that opens a card then to call that function within a second function.  Initially created a function to assign classes .open and .show, then attempted to call open() within an event listener which did not work.  Settled on assigning .open and .show directly within the event listener.

TODO Status:

// TODO: CREATE FUNCTION TO OPEN CARD (DONE)
// TODO: SET UP EVENT LISTENER FOR CARD CLICK (DONE)
// TODO: CREATE LIST FOR OPEN CARDS (Current)
// TODO: CHECK LIST FOR MATCHING CARDS