code: Fix click count and restart
clickCount() was counting every single click including multiple clicks on a single element to determine the number of moves. Added a CSS styling to .open for `pointer-events: none`.  The restart button was not working properly causing the timer to count faster with each click of the restart button.  Instead of customizing each action resulting on click of restart I opted to reload the page upon clicking restart.